requirement.txt
streamlit>=1.35.0
moviepy>=1.0.3
python-dotenv>=1.0.1
openai>=1.30.1
pytube>=15.0.0
